api/requirements.txt
azure-functions==1.19.0
# Temporarily removed azure-storage-blob while debugging deployment timeouts.
